requirements.txt
# Stormline Management Bot Dependencies

python-telegram-bot==20.7
anthropic>=0.40.0
python-dotenv==1.0.1
aiohttp==3.13.3
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
ftfy>=6.0.0
